+++ AdventSolver/solutions/Day1Solution.h
@@ -23,8 +23,8 @@ class Day1Solution : public Solution {
 public:
     Day1Solution(const std::vector<std::string> & puzzleInput);
     std::string getTitle() const override;
-    int oneStarSolution() override;
-    int twoStarSolution() override;
+    long long oneStarSolution() override;
+    long long twoStarSolution() override;
 };
