+++ AdventSolver/solutions/Day3Solution.h
@@ -17,8 +17,8 @@ class Day3Solution : public Solution {
 public:
     Day3Solution(const std::vector<std::string> & puzzleInput);
     std::string getTitle() const override { return title; }
-    int oneStarSolution() override;
-    int twoStarSolution() override;
+    long long oneStarSolution() override;
+    long long twoStarSolution() override;
 };
