+++ AdventSolver/solutions/Day6Solution.h
@@ -68,8 +68,8 @@ class Day6Solution : public Solution {
 public:
     explicit Day6Solution(const std::vector<std::string> & puzzleInput);
     std::string getTitle() const override { return title; }
-    int oneStarSolution() override;
-    int twoStarSolution() override;
+    long long oneStarSolution() override;
+    long long twoStarSolution() override;
 };
